Make server loop a bit more resilient
The core of the problem is that my Transaction wrapper, when you drop it
without committing or rollback, it leaves the session with a transaction
open. I need to fix that in the wrapper code, which I need to fix in
Squiller, but for now, this will fix the symptom.