Do not allow negative votes after all
The math works fine, but it requires some code to sort into supporters
and detractors when displaying the votes. Even that is not too bad, but
then if somebody actually puts in a negative vote and it is displayed,
you don't make friends that way. Let's just not risk it.