Make transaction retries work in theory
But not in practice. I don't understand why and I don't want to debug it
right now. I had been running with one thread by accident all this time
and it works fine. Let's do that in production too then, for now. It
only takes ~1ms to handle the request anyway so load shouldn't be an
issue, though I'd like to see how long it takes on the underpowered VM.